refactor(ale_claw): Tier 4 — split prompt-report out of session.py
session.py aggregates four concerns; this peels off the smallest, most
self-contained one. build_system_prompt_report moves verbatim to the new
prompt_report.py (deps: json/time only). session.py re-exports it
(from .prompt_report import build_system_prompt_report) so every existing
`from .session import build_system_prompt_report` — __init__.py and the test —
keeps working unchanged. Pure move, no behavior change; all 3 import paths
resolve to the same object; full suite green (928 passed, 3 skipped).
